:recycle: isolate mocks in separate package to avoid test dependencies in final executable
