+++ internal/engines/lognotify/mocks/notifier.go
@@ -1,4 +1,4 @@
-package lognotify
+package mocks
 
 import (
 	"github.com/stretchr/testify/mock"
@@ -7,6 +7,7 @@ import (
 
 	"github.com/vladopajic/go-actor/actor"
 
+	"link-society.com/flowg/internal/engines/lognotify"
 	"link-society.com/flowg/internal/models"
 )
 
@@ -15,16 +16,16 @@ type MockNotifier struct {
 	mock.Mock
 }
 
-var _ LogNotifier = (*MockNotifier)(nil)
+var _ lognotify.LogNotifier = (*MockNotifier)(nil)
 
 // NewMockNotifier returns a LogNotifier whose calls can be stubbed and asserted.
-func NewMockNotifier() LogNotifier {
+func NewMockNotifier() lognotify.LogNotifier {
 	return &MockNotifier{}
 }
 
-func (m *MockNotifier) Subscribe(ctx context.Context, stream string) (actor.MailboxReceiver[LogMessage], error) {
+func (m *MockNotifier) Subscribe(ctx context.Context, stream string) (actor.MailboxReceiver[lognotify.LogMessage], error) {
 	args := m.Called(ctx, stream)
-	return args.Get(0).(actor.MailboxReceiver[LogMessage]), args.Error(1)
+	return args.Get(0).(actor.MailboxReceiver[lognotify.LogMessage]), args.Error(1)
 }
 
 func (m *MockNotifier) Notify(ctx context.Context, stream string, logKey string, logRecord models.LogRecord) error {
